PA21-027 Register more entities
Some mapping between Entity_Name and Entity_Id was missing. This was
because we were not registering Entity_Ids in a number of cases.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Generated_Calls): we do not need to check if Entity_Id is present because
there should always be an Entity_Id if we are in this code.

* gnat2why/spark/spark_register.adb
(Process_Node): we now register every subprogram call. We also register
calls coming from type declarations, in particular any subprogram call in
predicates or DIC.

Change-Id: Icaa57939dc9558c8daae0917275924766f39534e

Former-commit-id: aa005ea6abcdc3331606e38a0255712276632e44