PB02-030 flow - detect side effect in PO
Functions must not have an input with effective reads as this would hide a
side-effect. This was already detected for global inputs of this kind. We
now issue an error message also in the case of protected functions that use
a volatile variable with Effective_Reads and Part_Of (which is not
considered to be a global input).

* gnat2why/flow/flow-analysis-sanity.ads
* gnat2why/flow/flow-analysis-sanity.adb
(Check_Side_Effects_In_Protected_Functions): we check for hidden side
effects in protected functions. We detect when a protected function uses a
volatile variable with both Effective_Reads and Part_Of the PO and we issue
a message.

* testsuite/gnatprove/tests/OA14-015__volatile_state_full_vol
* testsuite/gnatprove/tests/P118-029__volatile_protected
* testsuite/gnatprove/tests/PB02-030__flow_functions_with_volatile_input
Update test outputs

Change-Id: Ic92e6fa5751addb225b2f26668d052db31275210

Former-commit-id: 387792b34e5a850e4797333eb00a87670c98d9eb