P518-021 clarify comment about Get_Loop_Writes
Comment that was added as part of this TN was misleading: is suggested that
Get_Loop_Writes returns variables that "may" (?) be declared locally in the
loop. This suggests that for a code like this:

  loop
     declare
        X : Boolean;
        pragma Unreferenced (X);
     begin
        null;
     end;
  end loop;

the Get_Loop_Writes routine will also return the local variable X. This is
not the case, because X variable is neither "written" or volatile and
"read".

* gnat2why/flow/flow_utility.ads
(Get_Loop_Writes): clarify comment.

Change-Id: I4c408978e15878224431b837b47ad29221700839

Former-commit-id: 80ea802bd2d04a0e9975aa5564a005bbced83636