PB10-015 refined contracts on stubs
Refined contracts on separate entities are attached to the stub entity, not
to the spec/body entity. Now this is taken into account, at least in
Find_Contract.

* gnat2why/spark/spark_util-subprograms.adb
(Find_Contract): support refined contracts on stubs.
(Find_Contracts): only mark the missing functionality with ???; this routine
 should most likely be merged with Find_Contract.

* testsuite/gnatprove/tests/M729-011__flow_body_stubs/test.out
Update test output; now the error message for subprogram with separate body
is attached to the refined contract; previously it was attached to the
non-refined one, which was a safe fallback.

Change-Id: Id8f56f9a3e2f49d82e94ff2cb4850b178b87e3a6

Former-commit-id: b550e6fb4658d50d7522a4922b1a1ecb47e5865a