P914-008 array ordering lemma
This commits adds a lemma on array ordering. This is proved in Coq and is
checkable via the tests for lemmas inside the testsuite.
It also contains scripts for maintanability of lemmas library.

* include/README
README to explain how to generate the lemmas sessions.
* include/generate_session.py
Small executable to regenerate the session from scratch.
* include/generate_sessions.sh
Removed this now obsolete file.
* include/manual_proof.in
File/line corresponding to the manual proofs in lemma library.
* include/Coq/common/*.prf
Minor updates to proof because of change in SPARK.v
* include/proof/Coq/common/lemma_raising_order_float.prf
Added proof.
* include/proof/Coq/common/lemma_raising_order_int.prf
Added proof.
* include/Proof/Coq/spark-test_array_lemmas/*.ctx
Added context.
* include/proof/sessions/*/*
Updated session.
* include/spark-*.adb
Changed preprocessing puting spark_mode on/off to a removing of spark
mode off in test of lemmas library.
* include/spark-constrained_array_lemmas.ad*
Lemmas for constrained array. Does not work yet.
* include/spark-test_array_lemmas.ad*
Test file inside the library because it is impossible to debug this into
the testsuite.
* include/spark-unconstrained_array_lemmas.ad*
Core library for arrays. Even constrained array proof will depend on this.
* testsuite/gnatprove/tests/P405-006__coq_lemma_library/test.py
Adding replace_spark_mode to replace the spark_mode=off line given to
customer by the spark_mode=on one to check it in spark testsuite.
* testsuite/gnatprove/tests/P405-006__coq_lemma_library/test.out
Updated output.
* testsuite/gnatprove/tests/P914-008__trivial_test/*
Adding trivial example to show ordering lemma work. The user raising this
ticket had a workaround for ages. So no other test were directly
available.

Change-Id: Ie60989ba216b858179066daf2b7d37cf92083457

Former-commit-id: 650a80bf60bb14684df7ed890424be1fb840bc04