P715-019 Termination for subprograms in generic predefined units
We now deal differently with subprograms in a generic predefined unit. In
particular we assume they are nonreturning only if the generic is
instantiated with nonreturning subprograms.

* gnat2why/flow/flow.adb
(Build_Graphs_For_Entity): We now register as direct calls also actuals
used in a generic instantiation whenever we find a subprogram in a
predefined generic unit. We also do not register as nonreturning
subprograms with no body yet.

* gnat2why/spark/spark_util-subprograms.ads
(Is_Intrinsic): declare a function that returns True if a subprogram is
intrinsic

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Iterate_Generic_Actual_Parameters): created generic procedure to iterate
over generic actual parameters.

* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination
Add few tests and updated output.

Change-Id: I473c314f3c5ca1de6ada1faef2d7a58997f82f13

Former-commit-id: dd5f7fc1d820d8fcb97c1be6d2db2df42fe8f141