simplify string (de)allocation when compiling Coq files
In the Compile_Lib routine the assignment "Source_Dest := null" had no
effect (maybe except for silencing the compiler). The management of memory
pointed to by this variable was over-complicated: the pointer was copied to
Args (4) and Free was called only with this copy, never with the original.

This patch cleans this up.

* gnatprove/configuration.adb
(Compile_Lib): cleanup handling of Source_Dest.

(no-tn-check)

Change-Id: Ia8b9145a7356d60f4640eb18763fbef9b5e6a050

Former-commit-id: 4e2b1862843f77547261354c8d0847d1d344419c