PB01-009 re-enable test
Finding contracts on single concurrent objects is not fixed, but I forgot to
re-enable one test where this did not work.

This test was XFAIL-ed with the TN related to original work on contracts for
tasking. This was a mistake. It should have been XFAIL-ed with the TN
related to fixing contracts on single concurrent type.

* testsuite/gnatprove/tests/OA23-011__flow_garbage_tk_suffix/test.opt
Remove to re-enable test.

Change-Id: I0005f2a5943074b7cfa7d8b517648a07029430a4

Former-commit-id: 49644b57706fca22ba5b5933a0d128cc50c7b391