PB10-014 proof - update CVC4 and Z3.
Updated CVC4 and Z3. Fixes re-definition of Boolean in Why3.

Large update to testsuite testsuite. Pretty boring, except in the
following cases:

Fixes Z3 soundness bugs in:
   * KA11-014__fixed_points
   * L402-021__float

Added proof replays to a few tests.

Use Z3_noquant in the proofinuse tests. Also fix an incorrect annotation
there (the test is SAT, but not in the place indicated).

Use custom flags for bitwalker test in order to meet the no_fail
annotation.

Change-Id: I9797674088ebabe00c87d70018ffa5b0ee0788de

Former-commit-id: f12ed9f06fd539c006afdb1c49a76f3782c31d36