no calls to elaboration of nested wrapper packages
Do not register calls to nested wrapper packages, because they are not
flow-analyzed and we do not get direct calls for them; in effect, we can't
resolve calls from their elaboration (but there shouldn't be any such calls
anyway).

In old GG (phase 2) direct calls for wrapper packages are not registered in
the ALI file and we default to empty calls.

Needed for the new GG where we have much more assertions.

(no-tn-check)

Change-Id: Ifc5852e393e797d6b66cf3a1912a9bae75a7a37d

Former-commit-id: b3639ff19c8f7e80449cc9996a6a5afa7b3165f7