P715-007 No axiom for potentially nonreturning functions
Axioms generated for functions by GNATprove may be incorrect if the
function does not terminate on some of its input. If flow analysis
cannot make sure that a function always terminate, do not generate
an axiom for it.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post): Do not generate an axiom if the function
is potentially nonreturning.
(Translate_Expression_Function_Body): For functions potentially
nonreturning, generate the safe version of the expression function
defining axiom, with appropriate guard and conversions.

* gnat2why/why/gnat2why-annotate.adb
(Check_Terminate_Annotation): Also accept a pragma Annotate
Terminating on a package, the pragma applies to every subprogram
declaration in the package.
(Has_Terminate_Annotation): If the subprogram is not itself
annotated with a Terminating pragma, look for the pragma on the
enclosing packages if any.

Change-Id: I73445e088ff9c772231d21072506155d43633396

Former-commit-id: 0c335a3033d7f0f3b79e7420764de2be596e2c67