P907-030 Improve error message
In case of a subprogram having an output global which is used as an input
of the subprogram in its body we now provide more information on the error
message.

* gnat2why/flow/flow-analysis.adb:
(Find_Use_Of_Uninitialized_Variables): improve error message in the case
described above.

* testsuite/gnatprove/tests/
Update several test outputs with extended error message

Change-Id: I791202aef21f2f9f722f52581c997f61d47746e3

Former-commit-id: 03d12914de93ac998c75e92f74e2460c43b8ec8b