PA30-001 protected objects with components out of SPARK
Support protected objects with a full view not in SPARK. They can
arise either because their private part is explicitey marked Off or
because they are declared in a part with no SPARK_Mode but have
components which turn out to be out of SPARK.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): If the private part of a concurrent type is
marked Off, store it in Full_View_Not_In_SPARK. Do the same if the
mode is Auto and a violation is detected while marking the
components.

* gnat2why/spark/spark_util-types.adb
(Get_Variant_Info): Handle concurrent types with full view not in
SPARK.
(Has_Private_Fields): idem.
(Init_Component_Info_For_Protected_Types): idem.
(Init_Component_Info_For_Subtypes): New function that initialized
component information for subtypes of record and concurrent types.

* gnat2why/spark/spark_util.adb
(Component_Is_Visible_In_SPARK): Handle protected components of a
type with full view not in SPARK.

Change-Id: Ibef5536188f87cc2a684a7727e1a56293c154fa5