PB24-017 separate marking of protected type entity / declaration
When marking a protected type entity, we were traversing all its
declarations. We now separate marking of the type and marking of
the declarations so we are closer with the vision of a protected
type being a type and a set of operations. As a result, all
protected operartions will not be pulled whenever the protected
type is pulled anymore. They will be marked when needed, or right
after the type entity if SPARK_Mode is On.

* gnat2why/spark/spark_definition.adb
(Mark_Concurrent_Type_Declaration): New procedure to mark
declarations occuring inside protected types.
(Mark_Aspect_Clauses_And_Pragmas_In_List): New procedure that
traverses a list of nodes and only marks pragmas and aspect
clauses. It is used to mark pragmas associated to a protected
type while marking the type entity.
(Mark): Mark protected type declarations.
(Mark_Type_Entity): Do not mark declarations of protected
subprograms as part of marking protected types anymore.
(Mark_Entity): Use the correct SPARK_Mode when marking a
subprogram entity.

Change-Id: I7463ff9c4e441e2894d7287627d8195f9b333f7f