PB30-023 Assume value of string literals
String literals were translated as an abstract logic symbol with no
axiom. As a result, GNATprove could not reason about the content of
a string literal. We now generate an axiom that describes the
content of the logic symbol.
If the literal is translated as a constant, the axioms slows
GNATprove when a lot of strings are defined. To avoid this, we
translate them as functions with a unit parameter so that the axiom
can be protected by a trigger.

* gnat2why/why/gnat2why-expr.adb
(Transform_String_Literal): Loop over the values of the string
literal to generate an axiom describing the content of the why
map corresponding to the literal.
(Transform_Expr): Now the logic symbol for the literal is a
function with a unit parameter instead of a constant. Handle it
appropriately.

Change-Id: Ifc96bf49ec4e7def7b27abdb3ecfb294c27d14ed