PC01-022 more restrictive test for abstract Part_Of concurrent objects
Previous code had to test 3 conditions before a violation was raised; with
this patch there are only 2 tests. (The 3rd of the old tests was needed for
abstract states that are Part_Of other abstract states, which was giving
them a E_Void type. It was a bit confusing that we need to care about nested
and non-nested abstract states here.)

Also, in the old code the E_Protected_Subtype and E_Task_Subtype (which are
part of Protected_Kind and Task_Kind, respectively) case branches were never
executed (because Part_Of cannot be related to objects whose type is
E_..._Subtype).

Finally, now we avoid repeating the common prefix of the error message.

* gnat2why/spark/spark_definition.adb
(Mark_Entity): small rewrites.

Change-Id: Iac6779c6282bcae3f7d76d7b633160abcef14e51