PB16-052 Use intermediate assertions to prove complex float properties
By bounding float quantities by constants, CVC4 manages to prove easily
properties related to monotonocity of operations on floats. The assertions
on bounds are easily proved by CodePeer. Currently Z3 does not do much
on this example. The only remaining unrpoved assertion on attemp_3
involves integer to float conversions, and should be provable with a
lemma, as this is in the range where operations on integers and floats
give the same results (these integers are representable in floats).

Change-Id: Id32ba346839de4c775f934b14b62f8150dfbe4f0