PC05-011 non-terminating elaboration of nested packages
Elaboration of nested packages is now represented as a subprogram call;
there is no need to pretend that the non-terminating statements (e.g.
non-terminating loops or calls to predefined procedures with No_Return)
occur in the enclosing subprogram.

The original was P715-019 (subprogram termination)
Relies on P929-043 (elaboration of nested packages as subprogram calls)

Also, this gives a simpler code and is in line with the new GG approach.

* gnat2why/flow/flow.adb
(Flow_Analyse_Entity): do not pull register non-terminating nested packages
 under the name of their enclosing subprogram; this will be resolved later,
 in phase 2.

* testsuite/gnatprove/tests/PC05-011__flow_nested_nonterminating_pkgs/
Enable test; no longer crashes.

* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/test.out
Update with an incorrect test output that should be fixed by PB25-025.

Change-Id: If8f1daf63a2e16a53981be274cb9d35cd01bd41f