PB25-025 Improve termination
In case of a subprogram (annotated with the Terminating annotation) that
is nonreturning because it calls a nonreturning subprogram we were not
emitting a check. We are now.

* gnat2why/flow/flow-analysis.adb
(Check_Termination): if a check has been emitted on a subprogram when we
are not able to verify the Terminating annotation, do not emit an info
message on the termination status of the subprogram.

* gnat2why/flow/flow_generated_globals-phase_2.ads
* gnat2why/flow/flow_generated_globals-phase_2.adb
(Is_Potentially_Nonreturning_Internal): this is a renaming of the previous
Is_Potentially_Nonreturning. We now dive into each callee to check their
termination status even if the caller is annotated with the Terminating
annotation. This is used to decide when to emit a check if we cannot
verify the Terminating annotation.

(Is_Potentially_Nonreturning): returns True when
Is_Potentially_Nonreturning is True and does not have the Terminating
annotation.

* gnat2why/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post): re-use the new Is_Potentially_Nonreturning

(Translate_Expression_Function_body): re-use the new
Is_Potentially_Nonreturning

* testsuite/gnatprove/tests/PB25-025__flow_termination
Test added

* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination
Update test output

Change-Id: Idfe02baea4b56ffcd1afe2be377f4d9865434935