PA27-015 Fix message on call to non-returning procedure inside another
Call to a non-returning procedure inside another non-returning procedure
should be treated like a call to a returning procedure in terms of
message issued for a failing precondition. There was a missing detection
of this case in the message handling code, now fixed.

* gnat2why/why/gnat2why-error_messages.adb
(Error_Messages): Adjust message for precondition failure on
  non-returning procedure inside another non-returning procedure.

* testsuite/gnatprove/tests/PA27-015__nested_no_return/test.out
Test output update.

Change-Id: I34fa1e74ad93bfb59a7d1d753c4c99bb441c30a7