PC06-016 Range check on discrimiant of private type
When declaring a subtype or derived type of a private type with
full view not in SPARK, we should check that the supplied
constraint is in the range of the discrimiant type. This is now
done. For derived types, we can get two checks (the first one
potentially failing and the second one redundant) because we have
two declarations, one for the partial view, and one for the full
view.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Generate a range check for subtypes and
derived declarations of private types.

Change-Id: I9f48c91e3ab7c82258e52fae002ab36e98ec3d3c