PC06-030 Do not check record components of ancestors
When checking a record type declaration, we currently check subtype
indications on all scalar components, inherited or not. If the
components are inherited, we should not recheck them as we may lack
information which would make the checks unprovable.

* gnat2why/why/gnat2why-expr.adb
(Transform_Declaration): Do not introduce range checks for subtype
indications of record components if the declared type is not the
type in which the components were originally declared.

Change-Id: Ie35b7e02fdd2c510cb6dfd37b6c622cd60e5b3dd