PB25-025 exclude predefined subprograms from termination graphs
Calls to non-returning predefined subprograms are already detected in
phase-1 and effectively give the caller the status of non-returning. There
is no need to include them as dummy leafs in the call graph in phase-2.

This patch makes the call graphs slightly smaller and removes an ill-defined
call to Terminating_Subprograms.Contains (which is the primary reason as it
makes this code easier to adapt for the new GG).

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Is_Predefined): lift to library-level to allow reusing; now accepts
 Entity_Name and internally converts it to a string.

(Add_Termination_Edges): do not include predefined subprograms as dummy
 leafs in the termination call graph.

Change-Id: Ib944380b6f4a2ad74147173c179f0dcab086f296