PC08-035 Fix missing range check on argument of inlined call
When the actual parameter and the formal parameter of a call are
subtypes of a same scalar type with different constraints, and the call
is inlined because the procedure is local without contracts, GNATprove
can miss a range check on an assignment to the parameter. Now fixed.

* gnat2why/why/gnat2why-expr.adb
(New_Assignment): Deal with left-hand side of assignment that have
  range checks, as a result of inlining.
(Transform_Assignment_Statement): Do not introduce conversion here,
  instead let the call to New_Assignment introduce the conversion where
  needed so that range checks are properly inserted.

* gnat2why/why/why-gen-expr.adb
* gnat2why/why/why-gen-expr.ads
(Insert_Checked_Conversion): Add parameter Lvalue for checks on left-hand
  side of assignment, passed on to Insert_Scalar_Conversion.
(Insert_Scalar_Conversion): Add parameter Lvalue for checks on left-hand
  side of assignment, to decide on type for range checking. Do not apply
  optimization on floating-point ranges to left-hand side of assignment
  where Determine_Range_R is not suitable.

* testsuite/gnatprove/tests/P920-005__inline/test.out
Fix correct output.

* gnatprove/tests/PC08-035__missing_range_check_on_out_param/test.opt
* gnatprove/tests/PC08-035__missing_range_check_on_out_param/test.out
Test passes now.

Change-Id: Ibda1576b2accb845a88c0beec7f69ae519e6387a