PC09-010 classwide contracts on overridden inherited subprograms
We currently do not check inherited classwide contracts on
subprograms which override an inherited operation. This is due to
the fact that the contract nodes are not localized on the alias
introduced for the inherited functionsubprogram.
If the overridden subprogram is an alias, go to its ultimate alias
to find the contract.

* gnat2why/spark/spark_util-subprograms.adb
(Filter_Classwide_Contracts): Go to the ultimate alias of the
overriden subprogram.

Change-Id: I7e535ed9b10952b474463cf7500178b3ff1bd2cf