PC19-034 add test for suspension object passed as parameters
This test demonstrates what happens if multiple task call a procedure that
suspends on a suspension object given as a procedure parameter.

Note: it is not possible to call such a procedure with actual that is a
component of an array/record, because it breaks some Ada volatility rules
(RM C.6(12)).

* testsuite/gnatprove/tests/PC19-034__flow_suspension_actual/
Add test.

Change-Id: I7d416344e6912c6e36d346a3e0a0c98440dd87c3