PC21-009 rewrite generation of Initializes aspect
Generation of Initializes aspect relied on a ghost status that in phase-2 is
only known for variables from the withed compilation units. For other
variables the code was meaningless - perhaps it over-approximated or perhaps
the result was not really needed. I cannot tell what this code was really
doing.

This rewrite is needed to have a sane reference for the new GG.

Change-Id: I6fc3f2d58ac7de4d6be7876c94fcf3b33f054f66