PC07-016 Add test
The output complains about tasks TO1 and TO2 to access the same entry but
they are not because RO1.C1 and RO1.C2 are two different objects.
This will be fixed with this TN. At the moment we have --mode=flow because
we are currently crashing in proof on entry calls with composite prefix
(see PC09-008).

Change-Id: Ia35215dd60e5015356a61a0b704fa644a3db5239