PC07-016 Improve writing objects to ALI file
We used to write to the ALI file a synthetic name entry calls with
composite prefix. We now register the entire name.

* gnat2why/flow/flow-control_flow_graph.adb
(Do_Call_Statement): use Spark_Util.Full_Entry_Name to store the whole
prefix for an entry call.

* gnat2why/flow/flow.ads
Modified one component of the Entry_Call record to store an Entity_Name
instead of an Entity_Id.

* gnat2why/flow/flow_generated_globals-phase_1.adb
(Write_To_ALI): when registering tasking info to the ALI file we now
register the entire name of an object in the above sense.

* gnat2why/spark/spark_util.ads
* gnat2why/spark/spark_util.adb
(Full_Entry_Name): function to get the entire entry call name with the
whole prefix.

* testsuite/gnatprove/tests/PC07-016__flow_write_object_to_ali
Output is now updated with the correct one

Change-Id: I632559ba02e69e6d7a4be9dcfc1d9808cf1c1928