P712-014 Add Japanese version of core SPARK UG
This translation of the core part of SPARK User's Guide was created by
Mr. Masao Ito. The only changes with respect to his translation are
the changes to relative paths to add the 'ja' subdirectory, and labels
which are all prefixed with 'ja' to avoid conflicts with the English
version of the User's Guide, as both are maintained in the same project.

Change-Id: I3321497b98c9c4ba0064005fe8e8f837fd6129ee