P712-014 Fix references to non-existing labels and files in Japanese
The Japanese version was referring to files that did not exist anymore
(removed in commit a715c91), hence deleting this part of the explanations.
Also fix 30 references to labels which do not exist in the Japanese
version of SPARK UG, as this is the core only. For many references,
simply delete the parenthesis pointing to further information which does
not exist. In a few cases where the name of the reference is used as a
part of the explanatory text, simply remove the :ref: marker and quotes
and leave the English name of the label.

These fixes are needed to be able to compile the SPARK UG without
warnings (which is important as warnings are treated as errors with -W).

Change-Id: I6d5bfbae8275edf65e836d3ca2ab8b1894fe71d5