Q104-016 Remove spurious warnings on unused pragma Annotate for checks
When pragma Annotate is used to justify a proof check in a subprogram
that is inlined for proof, then a spurious warning is issued, reporting
that pragma Annotate does not justify any unproved check. Now fixed.

* gnat2why/why/gnat2why-annotate.adb
(Generate_Useless_Pragma_Annotate_Warnings): Do not issue the warning
  inside subprograms which are not requested for proof, in particular
  those that are always inlined for proof.

* testsuite/gnatprove/tests/Q104-016__annotate_inline/p.adb
* testsuite/gnatprove/tests/Q104-016__annotate_inline/test.out
Update test to use pragma Annotate on the statement rather than on
the inlined subprogram body.

Change-Id: I8a52a4f84e8214816d8c3582510da9d76b448eaf