Q105-001 check default values of hidden components
When a record object is default initialized, we need to check that
no runtime errors can occur in the default initialization of its
components. If the record is derived from a private type, we were
not checking the default values of its hidden components.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): If a type is tagged, its full view may be in
SPARK even if it is derived from a type whose full view is not in
SPARK. On such types, still set Full_View_Not_In_SPARK to True but
use the full view as representative to see the fields.

* gnat2why/spark/spark_util.adb
(Component_Is_Visible_In_SPARK): Components of a type with full
view not in SPARK which is not visibly derived from another type
cannot be in SPARK. The special case is needed so that components
visible in ancestor do not leak in completely hidden types.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Check): Also introduce checks for hidden
components. As we cannot tell the difference between components
hidden by a private ancestors (which we need to check) and
components hidden by a static discriminant (which we should not
check), we check both. Note that checks introduced in the second
case should be trivially true.

Change-Id: Ie7aa53d43d81f075ff8c3b3d8d9f7397f5e4b2e5