Q106-022 check Default_Value / Default_Component_Value of subtypes
Default values of scalar types and default component values of
array types are only stored in base types. Go to the base type to
search for default values and default component values.

* gnat2why/spark/spark_util-types.adb
(Base_Retysp): Go through private types to find the actual base
type of a subtype.

* gnat2why/why/gnat2why-expr.adb
(Compute_Default_Check): Go to the base type to find the default
value / default component value of the type.
(Compute_Default_Init): Go to the base type to find the default
value / default component value of the type.
(Variables_In_Default_Init): Go to the base type to find the
default value / default component value of the type.

* gnat2why/why/why-gen-expr.adb
(Get_Range_Check_Info): Go to the Retysp to find the proper type to
check.

Change-Id: I31fb17679d0611ae7fbe84a25d3a99d3412c4aaa