OC13-004 use Retysp to check if conversions are needed
When converting between two types, we check whether the target type
is the parent / base type of the source type to avoid introducing
checks in this case. If the source type is a private type, we
should compare to its representative type instead.

* gnat2why/spark/spark_util-types.adb
(Check_Needed_On_Conversion): Go to the representative type to
check if the target is the parent type / base type of the source
type.

Change-Id: I2e2fc9b6e909174b6a127939b6ee2458e48ec3bd