Q105-001 fix issues with full view not in SPARK
Avoid poluting the table containing first ancestors of types with a
full view not in SPARK with entities not in SPARK.

* gnat2why/spark/spark_definition.adb
(Mark_Type_Entity): Do not set the table containing first ancestors
of types with a full view not in SPARK if a violation was detected.

* gnat2why/spark/spark_util.adb
(Component_Is_Visible_In_SPARK): Check that the type is in SPARK
before querrying for its full view status.

Change-Id: Ide40bf75811f10594e910550e68c000f1d754907