PA12-001 check default values of private types at declaration
We used to check default values of types at use. For private types,
this unmodular behavior is counter intuitive. We now check default
values of private types only once, at declaration. As a result,
private types defined in library units are now verified as part of
the library, and not when the library is used.
Since default values of private types may depend on variables, a
new why module is introduced to generate VCs for them, so that they
do not depend on the values of variables at the type declaration
point.
The checks introduced for DIC depend on whether or not the DIC
expression depends on the current type instance. If it depends on
the type instance, it is considered as a postcondtion of the
default initialization of the private type, and is checked at
declaration. If it does not depend on the type instance, it is
considered as a precondition of the default initialization, and is
checked at use.

* gnat2why/spark/spark_definition.adb
(Get_SPARK_JSON): Add json status for private types.
(Mark_Type_Entity): Mark calls to primitives in DIC as unsupported.

* gnat2why/spark/spark_util-types.adb
(Get_Initial_DIC_Procedure): DIC procedures of derived types do not
have a body. Go to the type with the initial DIC declaration to get
the complete DIC procedure.
(Needs_DIC_Check): Removed.
(Needs_Default_Checks_At_Decl): Return True if the check for the
default initialization of a type should be done at declaration (for
specific private types without unknown dicriminants which do not
have a private full view).

* gnat2why/why/gnat2why-assumptions.adb
(Register_Proof_Claims): Do not search for postconditions if the
entity is a type.

* gnat2why/why/gnat2why-driver.adb
(Do_Generate_VCs): Generate VCs for type entities if they need
checks for default initialization at declaration.

* gnat2why/why/gnat2why-expr.adb
(Assume_Dynamic_Invariant_For_Variables): Function used to generate
assumptions for the dynamic invariants of a set of variables. It is
factorized out of Compute_Dynamic_Property_For_Inputs.
(Assume_Value_Of_Constants): Moved from gnat2why-subprograms.adb.
(Check_Type_With_DIC): New subprogram which inserts checks for the
default initial condition of a type.
(Check_Type_With_Invariants): Declaration moved to body. Only
assume DIC of the type variable if it is checked at use.
(Compute_Default_Check): New parameter Skip_Last_Cond to not check
the DIC of the current type if any. New parameter Include_Subtypes
to check for runtime errors of all subtypes of an unconstrained
type. New parameter New_Components to only check newly added
components in a tagged derivation. Only check DIC if it should be
checked at use. Only check public parts of types (do not go to
Retysp).
(Compute_Default_Init): Use Get_Initial_DIC_Procedure to find the
DIC expression if any.
(Variables_In_Default_Init): Use Get_Initial_DIC_Procedure to find
the DIC expression if any.

* gnat2why/why/gnat2why-subprograms.adb
(Assume_Value_Of_Constants): Moved to gnat2why-expr.adb.
(Compute_Dynamic_Property_For_Inputs): Call
Assume_Dynamic_Invariant_For_Variables.

* gnat2why/why/gnat2why-types.adb
(Generate_VCs_For_Type): New function to generate the VCs for a
type declaration. This module is used to check the default
initialization of private types.

* gnat2why/why/gnat2why-util.adb
(Check_DIC_At_Declaration): Return True if the DIC expression does
not reference the type instance.
(Needs_DIC_Check_At_Decl): Check if a type requires a DIC check at
declaration (it has its own DIC, the DIC is visible from SPARK and
the DIC expression references the type instance).
(Needs_DIC_Check_At_Use):  Check if a type requires a DIC check at
use (it has a DIC, inherited or not, visible from SPARK and the DIC
expression does not reference the type instance).

Change-Id: I211b58ebdcfb2bac66e5bdca8932c551bd3d2e13