O922-006 Classwide Pre and Post
In flow analysis we were not creating vertices for classwide Pre and Post
conditions and therefore issuing some wrong messages or not issuing some
others. We do now have vertices for them.

The output for N709-001__contracts it is not too clear at the moment,
maybe this will need fix in a separate TN.

* gnat2why/flow/flow_utility.adb
(Classwide_Pre_Post): functions that return the list of classwide pre or
postcondition (inherited or not) for an entity E.
(Get_Postcondition_Expressions): we now consider also classwide post
(Get_Precondition_Expressions): we now consider also classwide pre

* testsuite/gnatprove/tests/O922-006__flow_classwide_prepost
Add a test and updated output

* testsuite/gnatprove/tests/N709-001__contracts/test.out
* testsuite/gnatprove/tests/N709-001__dispatch/test.out
* testsuite/gnatprove/tests/NB04-027__classwide_prepost/test.out
* testsuite/gnatprove/tests/O217-033__stacks/test.out
* testsuite/gnatprove/tests/O311-007__wrong_lsp/test.out
* testsuite/gnatprove/tests/P325-014__dispatch/test.out
Test output updated

Change-Id: I57fcc3a6b5ede7a6126878f0f4c9ba43abfb8e98