minor update test toutput
The new test output looks a bit ugly, but it's just an extra warning.
For some reason, in this case, why3 thinks that Coq has crashed and
reports this fact to the user, copying the Coq error message. Not super
nice, but not a problem either. Leaving this as is now.

Change-Id: Ib8facbaf85003c9a9ce88733880d9e17eb399a3b