remove No_Effects from Flow_Analysis_Graphs record
Having No_Effects in Flow_Analysis_Graphs looks like a mistake. It can be
safely computed when needed and is never re-used. Also, negative variables
are confusing, it feels better to say when a routine *has* effects.

Note: "for all => not ..." iteration is replaced with "for some => ...",
which should have the same semantics (perhaps except on empty set, but it
does not really matter here) and is easier to read.

* gnat2why/flow/flow.ads
(Flow_Analysis_Graphs): remove No_Effects.

* gnat2why/flow/flow.adb
(Flow_Analyse_Entity): code for setting No_Effects is now refactored into
 Has_Effects.

(no-tn-check)

Change-Id: Iaba537ec7a1f849d165792e0c7d32d4e3a5c72f6