remove useless call to To_Entire_Variables and set union
This is semantic-preserving rewrite of this line to avoid useless work:

  Reads := To_Entire_Variables (Reads or Proof_Reads);

where Reads and Proof_Reads are disjoint (so it is more efficient to iterate
them separately) and To_Entire_Variables us useless (because both sets only
contain entire variables anyway).

Also, it makes debugging easier (and this is the main reason for this
rewrite): if we emit message then now it is clear where did the offending
global came from.

* gnat2why/flow/flow-analysis.adb
(Analyse_Main): rewrite.

(no-tn-check)

Change-Id: I812dfb23a262dc48433e963a52bd96f4aa544c1d