PA19-039 register entity names of dispatching operations
One of the design decisions for Register_Entity routine was to only register
between Entity_Id and Entity_Name for subprograms that are called. We did
this by searching for explicit call nodes in the AST.

However, this did not work for dispatching operations, which might be called
without explicit call nodes. Now we conservatively register all dispatching
operations. Alternative, i.e. trying to figure out when such an operation is
indeed called, is too complicated and not worth the effort.

Note: Register_Entity is already over-complicated and we repeatedly waste
time on this routine for very little gain in performance.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Generated_Calls): remove workaround, restore old version.

* gnat2why/spark/spark_register.adb
(Register_Entity): register dispatching operation where they are declared.

Change-Id: I53e5f9944b0c98ad4c1355b5717f8e40578efa2f