PC14-014 No generation of completion for types with external axioms
We used to generate a completion module for types declared in
packages with external axioms. It was necessary because these
modules could depend on the completion of generic parameters
of the package, which were not available in the generic
translation.
This mechanism is not needed anymore. Parts from the completion
should now be manually written in the provided translation.

* gnat2why/why/gnat2why-driver.adb
(Complete_Declaration): No need to complete declarations from
packages with external axioms.

* gnat2why/why/gnat2why-external_axioms.ads
(Complete_External_Entities): Removed.

* gnat2why/why/gnat2why-types.adb
(Generate_Type_Completion): Use better names for parameters of
dynamic invariant and default initialization predicates to improve
readability of the generated whyml code.

* gnat2why/why/why-atree-modules.adb
(E_Axiom_Module): There is only one module for entities declared in
packages with external axioms. Return this unique module when asked
for the axiom module of such an entity.

* docs/ug/gnatprove_by_example/examples/proof/_theories/sums.mlw
Add the parts comming from types axiom modules to their regular
modules.

Change-Id: Ib4cfa61d08e4e364c98319d1fe6a66e084350234