minor use Entity_Id-specific routines where possible
Routines that operate on Flow_Ids are handy when we don't know whether the
input is a direct mapping, magic string or something else. When processing
spec variables of a package the input must be a direct mapping, so there is
no need for a generic Flow_Id routine.

This patch clarifies that we only process direct mappings here and avoids
repeated conversion from Flow_Id to Entity_Id.

* gnat2why/flow/flow-analysis.adb
(Check_Elaborate_Body): use routines that operate on Entity_Id.

Change-Id: I8d7309a67ec975939179758d75bee91688d0148a