Q222-002 add test
It is tempting to use discriminants (either immediate or nested) in
expressions where variable input is not allowed, but such uses (either
direct or via a proxy function) are currently rejected.

The current behaviour is fine, but we didn't have a test for it. Here it is.

* gnat2why/flow/flow-analysis-sanity.adb
(Check_Flow_Id_Set): add comment about accepted discriminants.

* testsuite/gnatprove/tests/Q222-002__flow_nested_discriminant/
Add. Output is as expected. Only for coverage.

Change-Id: I7e8c06bea4406bef089ca5f0d0f1769c44472f0e